# Projeto_Compilador/api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Eu Digital API
  version: 1.0.0
  description: API RESTful para gestão do “eu digital” — ingestão de pacotes SIP, CRUD de fotos, textos, resultados académicos, resultados desportivos, ficheiros e eventos.
servers:
  - url: /api
paths:
...
  /photos:
    get:
      summary: List photos
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
        - in: query
          name: tag
          schema:
            type: string
          description: Filter by tag
        - in: query
          name: author
          schema:
            type: string
          description: Filter by author
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photo'
    post:
      summary: Create a photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Validation error
  /photos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: Photo UUID
    get:
      summary: Get a photo by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '404':
          description: Not found
    put:
      summary: Update a photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete a photo
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
  /texts:
    get:
      summary: List texts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: author
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Text'
    post:
      summary: Create a text entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
        '400':
          description: Validation error
  /texts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a text by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
        '404':
          description: Not found
    put:
      summary: Update a text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete a text
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
...
  /sportResults:
    get:
      summary: List sport results
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: activity
          schema:
            type: string
        - in: query
          name: author
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SportResult'
    post:
      summary: Create a sport result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SportResultInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportResult'
        '400':
          description: Validation error
  /sportResults/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get sport result by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportResult'
        '404':
          description: Not found
    put:
      summary: Update a sport result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SportResultInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportResult'
        '400':
          description: Validation error
        '404':
          description: Not found
    delete:
      summary: Delete a sport result
      responses:
        '204':
          description: No Content
        '404':
          description: Not found
...
components:
  schemas:
    BaseItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        author:
          type: string
        type:
          type: string
        visibility:
          type: string
          enum: [public, friends, private]
        tags:
          type: array
          items:
            type: string
      required: [id, author, type]
    Photo:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            format:
              type: string
            resolution:
              type: object
              properties:
                width:
                  type: integer
                height:
                  type: integer
            location:
              type: object
              properties:
                lat:
                  type: number
                lon:
                  type: number
                description:
                  type: string
            caption:
              type: string
    PhotoInput:
      allOf:
        - $ref: '#/components/schemas/Photo'
      required: [id, author, type]
    Text:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            title:
              type: string
            content:
              type: string
            summary:
              type: string
    TextInput:
      allOf:
        - $ref: '#/components/schemas/Text'
      required: [id, author, type, content]
...
    SportResult:
      allOf:
        - $ref: '#/components/schemas/BaseItem'
        - type: object
          properties:
            activity:
              type: string
            value:
              oneOf:
                - type: number
                - type: string
            unit:
              type: string
            location:
              type: string
            activityDate:
              type: string
              format: date-time
    SportResultInput:
      allOf:
        - $ref: '#/components/schemas/SportResult'
      required: [id, author, type, activity, value, unit, activityDate]
...